lambda_functions/aviationedge/airport_requests/requirements.txt
boto3
requests
python-dotenv
pyyaml